Update environment script

+++ Docker/scripts/set_env.sh
@@ -12,32 +12,52 @@ TOOLCHAIN_DIR=$WORK_DIR/x-tools
 
 #MAIN HANDLE
 case $1 in
-	"setup")
-		git clone https://github.com/crosstool-ng/crosstool-ng.git $CROSSTOOL_DIR
-		cd $CROSSTOOL_DIR
-		git checkout crosstool-ng-1.24.0
-		rm -rf ~/src
-		mkdir ~/src &>/dev/null
-		wget -P ~/src https://github.com/libexpat/libexpat/releases/download/R_2_2_6/expat-2.2.6.tar.bz2
-		wget -P ~/src https://libisl.sourceforge.io/isl-0.20.tar.gz
-		./bootstrap
-		./configure --prefix=$(pwd)
-		make
-		make install
-		cd $WORK_DIR
+	"setup_ct")
+		if [[ ! -d $CROSSTOOL_DIR]]; then
+			git clone https://github.com/crosstool-ng/crosstool-ng.git $CROSSTOOL_DIR
+			cd $CROSSTOOL_DIR
+			git checkout crosstool-ng-1.24.0
+			rm -rf ~/src
+			mkdir ~/src &>/dev/null
+			wget -P -nv ~/src https://github.com/libexpat/libexpat/releases/download/R_2_2_6/expat-2.2.6.tar.bz2
+			wget -P -nv ~/src https://libisl.sourceforge.io/isl-0.20.tar.gz
+			./bootstrap
+			./configure --prefix=$(pwd)
+			make
+			make install
+			cd $WORK_DIR
+		else
+			echo "CROSSTOOL is already installed"
+		fi
 		;;
-	"rpi")
-		BIN_DIR=$TOOLCHAIN_DIR/aarch64-rpi4-linux-gnu/bin
-		export CROSS_COMPILE=aarch64-rpi4-linux-gnu-
-		export ARCH=arm
-		;;
-	"qemu")
+	"setup_qemu")
+		append_path $CROSSTOOL_DIR
 		BIN_DIR=$TOOLCHAIN_DIR/arm-unknown-linux-gnueabi/bin
-		export CROSS_COMPILE=arm-unknown-linux-gnueabi-
-		export ARCH=arm
+		if [[ -d $CROSSTOOL_DIR ]]; then
+			if [[ ! -d $BIN_DIR ]]; then
+				cd $CROSSTOOL_DIR
+				bin/ct-ng distclean
+				bin/ct-ng arm-unknown-linux-gnueabi
+				sed -i 's/CT_PREFIX_DIR_RO=y/# CT_PREFIX_DIR_RO is not set/g' .config
+				bin/ct-ng build
+				cd $WORK_DIR
+			else
+				echo "TOOLCHAIN already installed"
+			fi
+		else
+			echo "CROSSTOOL is not installed"
+		fi
+		;;
+	"set_qemu")
+		append_path $CROSSTOOL_DIR
+		append_path $BIN_DIR
+		BIN_DIR=$TOOLCHAIN_DIR/arm-unknown-linux-gnueabi/bin
+		if [[ -d $BIN_DIR ]]; then
+			export CROSS_COMPILE=arm-unknown-linux-gnueabi-
+			export ARCH=arm
+		else
+			echo "TOOLCHAIN is not installed"
+		fi
 		;;
 esac
 
-#PATH HANDLE
-append_path $CROSSTOOL_DIR
-append_path $BIN_DIR
